♻️ Refactor updater to use Tauri commands and restructure release endpoint
- Migrate window management from window.ts to Tauri invoke commands
- Move latest.json to i18n-aware path with lang parameter support
- Update updater endpoint URL to match new release structure
- Simplify upgrade window creation by delegating to Rust backend
