+++ apps/desktop/src-tauri/tauri.conf.json
@@ -20,7 +20,7 @@
       "active": true,
       "pubkey": "dW50cnVzdGVkIGNvbW1lbnQ6IG1pbmlzaWduIHB1YmxpYyBrZXk6IEI4MjJCQTgwM0EyNEUyMjIKUldRaTRpUTZnTG9pdUFKSEJIZ3JnMG5Oay8zYVZjWW9rWisya2x3T2JhY1JsbnFvYzhqeDFsUFEK",
       "endpoints": [
-        "https://typo.yuler.cc/latest.json"
+        "https://typo.yuler.cc/releases/latest/download/latest.json"
       ],
       "windows": {
         "installMode": "passive"
